requirements.txt
# Core Framework
flask==2.3.3
gunicorn==21.2.0

# Telegram Bot Integration
requests==2.31.0
python-telegram-bot==20.7

# Google Cloud & Sheets API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0

# Environment & Configuration
python-dotenv==1.0.0

# PDF Generation & Analytics
reportlab==4.0.7
matplotlib==3.8.2
numpy==1.24.3
pandas==2.0.3

# Security & Performance
werkzeug==2.3.7
jinja2==3.1.2

# Development Dependencies (optional)
# Uncomment for development environment
# pytest==7.4.0
# black==23.7.0
# flake8==6.0.0
# mypy==1.5.1
# pre-commit==3.3.3